+++ packages/fleury_widgets/lib/src/approval_prompt.dart
@@ -1,6 +1,5 @@
 import 'package:fleury/fleury_core.dart';
 
-import 'controls.dart' show Button, ButtonVariant;
 import 'dialog.dart' show Dialog;
 import 'semantic_roles.dart';